Add Terraform configuration for S3 bucket

+++ main.py
@@ -1,60 +1,48 @@
 import boto3
+import subprocess
 import os
-from botocore.exceptions import ClientError
-from datetime import datetime, timezone, timedelta
+
+def get_bucket_name():
+    result = subprocess.run(
+        ["terraform", "output", "-raw", "bucket_name"],
+        cwd="./terraform",
+        capture_output=True,
+        text=True
+    )
+    return result.stdout.strip()
+
+bucket = get_bucket_name()
 
 s3_client = boto3.client('s3')
 
-def create_bucket(bucket_name):
-    try:
-        s3_client.head_bucket(Bucket=bucket_name)
-        print(f"Bucket '{bucket_name}' already exists.")
+def upload_files(bucket_name):
+    s3_client = boto3.client('s3')
+    upload_folder = 'test_uploads'
 
-    except ClientError:
-        region = s3_client.meta.region_name
-        if region == "us-east-1":
-            s3_client.create_bucket(Bucket=bucket_name)
-        else:
-            s3_client.create_bucket(
-                Bucket=bucket_name,
-                CreateBucketConfiguration={'LocationConstraint': region}
-            )
-        print(f"Bucket '{bucket_name}' created ssuccessfully!")
+    if not os.path.exists(upload_folder):
+        print(f"Folder '{upload_folder}' does not exist.")
+        return
 
-def upload_files(bucket_name, folder_path):
-    for file_name in os.listdir(folder_path):
-        file_path = os.path.join(folder_path, file_name)
+    for file_name in os.listdir(upload_folder):
+        file_path = os.path.join(upload_folder, file_name)
+        
         if os.path.isfile(file_path):
-            s3_client.upload_file(file_path, bucket_name, file_name)
-            print(f"Uploaded {file_name} to {bucket_name}")
+            try:
+                s3_client.upload_file(file_path, bucket_name, file_name)
+                print(f"Uploaded {file_name} to {bucket_name}")
+            except Exception as e:
+                print(f"Error uploading {file_name}: {e}")
 
 def list_files(bucket_name):
     response = s3_client.list_objects_v2(Bucket=bucket_name)
-    contents = response.get('Contents', [])
-    if not contents:
-        print("No files in the bucket.")
-    else:
-        for obj in contents:
-            print(obj['Key'])
-
-def cleanup_old_files(bucket_name, days_old):
-    cuttoff_date = datetime.now(timezone.utc) - timedelta(days=days_old)
-    response = s3_client.list_objects_v2(Bucket=bucket_name)
     if 'Contents' in response:
         for obj in response['Contents']:
-            last_modified = obj['LastModified']
-            if last_modified < cuttoff_date:
-                s3_client.delete_object(Bucket=bucket_name, Key=obj['Key'])
-                print(f"Deleted {obj['Key']} (last modified: {last_modified})")
-    
+            print(obj['Key'])
     else:
-        print("No files found for cleanup")
+        print("No files found.")
 
 
 if __name__ == "__main__":
-    bucket = "my-s3-cleaner-demo-bucket-2025"
-    folder = "test_uploads"
-    create_bucket(bucket)
-    upload_files(bucket, folder)
+    bucket = get_bucket_name()
+    upload_files(bucket)
     list_files(bucket)
-    cleanup_old_files(bucket, 3)
